requirements.txt
python-dotenv==1.0.0
bcrypt==4.0.1